split client tool handling into its own file
